add progress tracker

+++ vin_generate_tags.lua
@@ -173,17 +173,27 @@ end
     - The function currently only supports processing one image at a time.
 --]]
 local function generate_and_attach(images)
+    local job = dt.gui.create_job('Attaching tags and copyright to ' .. #images .. " images", true)
+    if #images ~= 1 then
+        dt.print("Please select exactly one image to generate tags and title.")
+        dt.print_hinter("Please select exactly one image to generate tags and title.")
+        return
+    end
     for _, img in ipairs(images) do
         dt.print_toast("Generating tags and title for: " .. (img.path) .. "/" .. (img.filename))
+        job.percent = 0
         local jpegfile = utils.convert_to_temp_jpg(img)
+        job.percent = 1 / 4
         dt.print_hinter("Generated temp file: " .. jpegfile)
         local tags, caption = generate_tags_title_description_with_ai(jpegfile)
+        job.percent = 3 / 4
         attach_tags_to_image(img, tags)
         add_title_to_image(img, caption)
+        job.percent = 1
         -- add_description_to_image(img, writeup)
         os.remove(jpegfile) -- clean up the temp file
-
     end
+    job.valid = false
 end
 
 -- defensive cleanup: remove any previous registration with the same name/type
